+++ app/src/lib/interfaces/ContextInterface/UiConfigurationProps.tsx
@@ -1,0 +1,43 @@
+export interface MojitoDeliveryType {
+  personalWallet?: boolean;
+  multiSig?: boolean;
+}
+
+export interface MojitoUIConfiguration {
+  global?: {
+    logoSrc?: string;
+    loaderImageSrc?: string;
+    errorImageSrc?: string;
+  };
+  billing?: {
+    isEnableExpressCheckout?: boolean;
+    gpay?: boolean;
+    applepay?: boolean;
+    walletConnect?: boolean;
+    metaMask?: boolean;
+  };
+  payment?: {
+    gpay?: boolean;
+    applepay?: boolean;
+    walletConnect?: boolean;
+    wire?: boolean;
+    creditCard?: boolean;
+    coinbase?: boolean;
+  };
+  costBreakdown?: {
+    showDiscountCode?: boolean;
+  };
+  paymentConfirmation?: {
+    wireTransferInstructions?: JSX.Element;
+    creditCardInstructions?: JSX.Element;
+    onGoTo?: () => void;
+  };
+  delivery?: {
+    gpay?: MojitoDeliveryType;
+    applepay?: MojitoDeliveryType;
+    walletConnect?: MojitoDeliveryType;
+    wire?: MojitoDeliveryType;
+    creditCard?: MojitoDeliveryType;
+    coinbase?: MojitoDeliveryType;
+  };
+}
