+++ app/src/lib/interfaces/ContextInterface/index.ts
@@ -2,3 +2,4 @@ export * from './CheckoutOptions';
 export * from './UIConfiguration';
 export * from './RootContainer';
 export * from './UserInfo';
+export * from './UiConfigurationProps';
